Extract useful methods from sqllogictest bin
